code/bot/requirements.txt
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
Flask==2.2.5
idna==3.4
importlib-metadata==4.13.0
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.1
MarkupSafe==2.1.3
nltk==3.8.1
numpy==1.21.6
psycopg2-binary==2.9.7
pyparsing==3.1.1
pyTelegramBotAPI==4.12.0
rdflib==6.3.2
regex==2023.6.3
requests==2.31.0
six==1.16.0
SPARQLWrapper==2.0.0
telebot==0.0.5
tqdm==4.65.0
typing_extensions==4.7.1
urllib3==2.0.4
Werkzeug==2.2.3
zipp==3.15.0
